refactor(gateway): inline update request merging
